pyproject.toml
[project]
name = "ais"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pydantic>=2.12.5",
    "typer>=0.24.1",
    "pandas>=3.0.1",
    "openpyxl>=3.1.5",
    "python-dotenv>=1.0.1",
    "rich>=14.3.3",
    "mypy>=1.11",
]

[dependency-groups]
dev = [
    "cuid2>=2.0.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "testcontainers>=4.14.1",
]
